First attempt at allowing non-vagrant installs

+++ .setup/install_system.sh
@@ -15,6 +15,7 @@ fi
 #################################################################
 # UBUNTU SETUP
 #################
+if VAGRANT; then 
 echo -e '
  __   __  _     _  _______  _______  ______    __   __  _______  ______
 |  | |  || | _ | ||       ||       ||    _ |  |  | |  ||       ||    _ |
@@ -42,6 +43,7 @@ chmod +rx /etc/motd
 echo "192.168.56.101    test-submit test-submit.cs.rpi.edu" >> /etc/hosts
 echo "192.168.56.102    test-svn test-svn.cs.rpi.edu" >> /etc/hosts
 echo "192.168.56.103    test-hwgrading test-hwgrading.cs.rpi.edu hwgrading" >> /etc/hosts
+fi
 
 #################################################################
 # PACKAGE SETUP
@@ -73,8 +75,7 @@ libseccomp2 seccomp junit cmake
 # Install Oracle 8 Non-Interactively
 echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | sudo /usr/bin/debconf-set-selections
 apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys EEA14886
-echo "apt-get install -qqy oracle-java8-installer..."
-apt-get install -qqy oracle-java8-installer > /dev/null 2>&1
+apt-get install -qqy oracle-java8-installer
 
 #################################################################
 # JAR SETUP
@@ -125,17 +126,20 @@ echo -e "#%PAM-1.0
 auth required pam_unix.so
 account required pam_unix.so" > /etc/pam.d/httpd
 
+if VAGRANT; then
 # Loosen password requirements
-sed -i '25s/^/\#/' /etc/pam.d/common-password
-sed -i '26s/pam_unix.so obscure use_authtok try_first_pass sha512/pam_unix.so obscure minlen=1 sha512/' /etc/pam.d/common-password
+	sed -i '25s/^/\#/' /etc/pam.d/common-password
+	sed -i '26s/pam_unix.so obscure use_authtok try_first_pass sha512/pam_unix.so obscure minlen=1 sha512/' /etc/pam.d/common-password
+# Set the ServerName
+	echo -e "\nServerName 10.0.2.15\n" >> /etc/apache2/apache2.conf
+fi
 
+# comment out directory configs - should be converted to something more flexible
 sed -i '153,174s/^/#/g' /etc/apache2/apache2.conf
 
 # remove default sites which would cause server to mess up
 rm /etc/apache2/sites*/000-default.conf
 rm /etc/apache2/sites*/default-ssl.conf
-
-echo -e "\nServerName 10.0.2.15\n" >> /etc/apache2/apache2.conf
 
 service apache2 reload
 
@@ -155,62 +159,52 @@ sed -i -e 's/^session.cookie_httponly =/session.cookie_httponly = 1/g' /etc/php5
 #################################################################
 # USERS SETUP
 #################
-adduser vagrant sudo
-
 addgroup hwcronphp
 addgroup course_builders
-adduser ta --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
-echo "ta:ta" | sudo chpasswd
-adduser instructor --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
-echo "instructor:instructor" | sudo chpasswd
-adduser instructor sudo
-adduser developer --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
-echo "developer:developer" | sudo chpasswd
-adduser developer sudo
+
+if VAGRANT; then
+	adduser vagrant sudo
+	adduser ta --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
+	echo "ta:ta" | sudo chpasswd
+	adduser instructor --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
+	echo "instructor:instructor" | sudo chpasswd
+	adduser instructor sudo
+	adduser developer --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
+	echo "developer:developer" | sudo chpasswd
+	adduser developer sudo
+fi
+
 adduser hwphp --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
-echo "hwphp:hwphp" | sudo chpasswd
+if VAGRANT; then
+	echo "hwphp:hwphp" | sudo chpasswd
+fi
 adduser hwcron --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
-echo "hwcron:hwcron" | sudo chpasswd
+if VAGRANT; then
+	echo "hwcron:hwcron" | sudo chpasswd
+fi
 adduser hsdbu --gecos "First Last,RoomNumber,WorkPhone,HomePhone" --disabled-password
-echo "hsdbu:hsdbu" | sudo chpasswd
+if VAGRANT; then
+	echo "hsdbu:hsdbu" | sudo chpasswd
+fi
 adduser hwphp hwcronphp
 adduser hwcron hwcronphp
 
-# TODO: we can automate this with a loop probably
-addgroup csci1100
-adduser ta csci1100
-adduser instructor csci1100
-adduser developer csci1100
-addgroup csci1100_tas_www
-adduser hwphp csci1100_tas_www
-adduser hwcron csci1100_tas_www
-adduser ta csci1100_tas_www
-adduser instructor csci1100_tas_www
-adduser developer csci1100_tas_www
+for COURSE in csci1100 csci1200  csci2600
+do
+	addgroup $COURSE
+	addgroup $COURSE_tas_www
+	adduser hwphp $COURSE_tas_www
+	adduser hwcron $COURSE_tas_www
+	if VAGRANT; then
+		adduser ta $COURSE
+		adduser instructor $COURSE
+		adduser developer $COURSE
+	fi
+done
 
-addgroup csci1200
-adduser ta csci1200
-adduser instructor csci1200
-adduser developer csci1200
-addgroup csci1200_tas_www
-adduser hwphp csci1200_tas_www
-adduser hwcron csci1200_tas_www
-adduser ta csci1200_tas_www
-adduser instructor csci1200_tas_www
-adduser developer csci1200_tas_www
-
-addgroup csci2600
-adduser ta csci2600
-adduser instructor csci2600
-adduser developer csci2600
-addgroup csci2600_tas_www
-adduser hwphp csci2600_tas_www
-adduser hwcron csci2600_tas_www
-adduser ta csci2600_tas_www
-adduser instructor csci2600_tas_www
-adduser developer csci2600_tas_www
-
-adduser instructor course_builders
+if VAGRANT; then
+	adduser instructor course_builders
+fi
 
 mkdir -p /var/local/hss
 mkdir -p /usr/local/hss
@@ -237,29 +231,33 @@ chmod g+s /var/lib/svn
 mkdir -p /var/lib/svn/csci2600
 touch /var/lib/svn/svngroups
 chown www-data:csci2600_tas_www /var/lib/svn/csci2600 /var/lib/svn/svngroups
-su hwcron
-echo -e "\n" | ssh-keygen -t rsa -b 4096 -N ""
-echo "hwcron" > password.txt
-sshpass -f password.txt ssh-copy-id hwcron@test-svn
-rm password.txt
-echo "csci2600_tas_www: hwcron ta instructor developer" >> /var/lib/svn/svngroups
+if VAGRANT; then
+	su hwcron
+	echo -e "\n" | ssh-keygen -t rsa -b 4096 -N ""
+	echo "hwcron" > password.txt
+	sshpass -f password.txt ssh-copy-id hwcron@test-svn
+	rm password.txt
+	echo "csci2600_tas_www: hwcron ta instructor developer" >> /var/lib/svn/svngroups
+fi
 
 service apache2 restart
 
 #################################################################
 # POSTGRES SETUP
 #################
-echo "postgres:postgres" | chpasswd postgres
-adduser postgres shadow
-service postgresql restart
-sed -i -e "s/# ----------------------------------/# ----------------------------------\nhostssl    all    all    192.168.56.0\/24    pam\nhost    all    all    192.168.56.0\/24    pam/" /etc/postgresql/9.3/main/pg_hba.conf
-echo "Creating PostgreSQL users"
-su postgres << EOF
-psql -c "
-  CREATE ROLE hsdbu WITH SUPERUSER CREATEDB CREATEROLE LOGIN PASSWORD 'hsdbu';
-  CREATE ROLE root WITH SUPERUSER CREATEDB CREATEROLE LOGIN PASSWORD 'vagrant';
-  CREATE ROLE vagrant WITH SUPERUSER CREATEDB CREATEROLE LOGIN PASSWORD 'vagrant';"
-EOF
+if VAGRANT; then
+	echo "postgres:postgres" | chpasswd postgres
+	adduser postgres shadow
+	service postgresql restart
+	sed -i -e "s/# ----------------------------------/# ----------------------------------\nhostssl    all    all    192.168.56.0\/24    pam\nhost    all    all    192.168.56.0\/24    pam/" /etc/postgresql/9.3/main/pg_hba.conf
+	echo "Creating PostgreSQL users"
+	su postgres << EOF
+	psql -c "
+	  CREATE ROLE hsdbu WITH SUPERUSER CREATEDB CREATEROLE LOGIN PASSWORD 'hsdbu';
+	  CREATE ROLE root WITH SUPERUSER CREATEDB CREATEROLE LOGIN PASSWORD 'vagrant';
+	  CREATE ROLE vagrant WITH SUPERUSER CREATEDB CREATEROLE LOGIN PASSWORD 'vagrant';"
+	EOF
+fi
 
 #################################################################
 # HWSERVER SETUP
@@ -276,24 +274,30 @@ fi
 HWSERVER_DIR=/usr/local/hss/GIT_CHECKOUT_HWserver
 cd ${HWSERVER_DIR}
 
-echo -e "localhost
+if VAGRANT; then
+	echo -e "localhost
 hsdbu
 hsdbu
 http://192.168.56.103
 svn+ssh:192.168.56.102" | source ${HWSERVER_DIR}/CONFIGURE.sh
+else
+	source ${HWSERVER_DIR}/CONFIGURE.sh
+fi
 
 source ${HWSERVER_DIR}/INSTALL.sh
 
 source ${HWSERVER_DIR}/Docs/sample_bin/admin_scripts_setup
-sed -i 's/SSLCertificateChainFile/#SSLCertificateChainFile/g' /root/bin/bottom.txt
-sed -i 's/course01/csci2600/g' /root/bin/gen.middle
-
 cp ${HWSERVER_DIR}/Docs/sample_apache_config /etc/apache2/sites-available/submit.conf
-sed -i 's/hss.crt/submit.crt/g' /etc/apache2/sites-available/submit.conf
-sed -i 's/hss.key/submit.key/g' /etc/apache2/sites-available/submit.conf
 cp ${HWSERVER_DIR}/Docs/hwgrading.conf /etc/apache2/sites-available/hwgrading.conf
-sed -i 's/SSLCertificateChainFile/#SSLCertificateChainFile/g' /etc/apache2/sites-available/hwgrading.conf
-sed -i 's/hwgrading.cer/hwgrading.crt/g' /etc/apache2/sites-available/hwgrading.conf
+
+if VAGRANT; then
+	sed -i 's/SSLCertificateChainFile/#SSLCertificateChainFile/g' /root/bin/bottom.txt
+	sed -i 's/course01/csci2600/g' /root/bin/gen.middle
+	sed -i 's/hss.crt/submit.crt/g' /etc/apache2/sites-available/submit.conf
+	sed -i 's/hss.key/submit.key/g' /etc/apache2/sites-available/submit.conf
+	sed -i 's/SSLCertificateChainFile/#SSLCertificateChainFile/g' /etc/apache2/sites-available/hwgrading.conf
+	sed -i 's/hwgrading.cer/hwgrading.crt/g' /etc/apache2/sites-available/hwgrading.conf
+fi
 
 a2ensite submit
 a2ensite hwgrading
@@ -308,7 +312,6 @@ if [[ ${VAGRANT} == 1 ]]; then
   rm /vagrant/.vagrant/tagrading_logs
   ln -s /vagrant/.vagrant/autograding_logs /var/local/hss/autograding_logs
   ln -s /vagrant/.vagrant/tagrading_logs /var/local/hss/tagrading_logs
-fi
 
 #################################################################
 # CRON SETUP
@@ -357,3 +360,4 @@ psql -d hss_csci1200_f15 -h localhost -U hsdbu -f ${HWSERVER_DIR}/TAGradingServe
 psql -d hss_csci1200_f15 -h localhost -U hsdbu -f ${HWSERVER_DIR}/TAGradingServer/data/inserts.sql
 psql -d hss_csci2600_f15 -h localhost -U hsdbu -f ${HWSERVER_DIR}/TAGradingServer/data/tables.sql
 psql -d hss_csci2600_f15 -h localhost -U hsdbu -f ${HWSERVER_DIR}/TAGradingServer/data/inserts.sql
+fi
